// src/@types/Dashboard.d.ts
export interface AppDashboardType {
    "Openlead": number,
    "ROlead": number,
    "Assignedlead": number,
    "ReAssigned": number,
    "RoConfirmation": number,
    "QC": number,
    "QCHold": number,
    "Pricing": number,
    "CompletedLeads": number,
    "OutofTATLeads": number,
    "DuplicateLeads": number,
    "PaymentRequest": number,
    "RejectedLeads": number
}